+++ src/dataIngestion/docker-compose.yaml
@@ -23,3 +23,15 @@ services:
       - text_files:/data/files
     depends_on:
       - whisper-setup
+
+    backend:
+      build:
+        context: ../backend/
+        dockerfile: .
+      volumes:
+        - whisper_model:/data/model
+        - extracted_audio:/data/audio/files
+        - text_files:/data/files
+      depends_on:
+        - whisper-setup
+        - multimedia-handler
